Move more to :compose and :util
