+++ app/src/main/java/de/schnettler/scrobbler/screens/local/ScrobbleItem.kt
@@ -32,12 +32,12 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import de.schnettler.database.models.Scrobble
 import de.schnettler.scrobbler.R
-import de.schnettler.scrobbler.ui.common.compose.NameListIcon
 import de.schnettler.scrobbler.ui.common.compose.CustomDivider
+import de.schnettler.scrobbler.ui.common.compose.NameListIcon
+import de.schnettler.scrobbler.ui.common.util.asMinSec
 import de.schnettler.scrobbler.ui.common.util.milliSecondsToDate
 import de.schnettler.scrobbler.ui.common.util.packageNameToAppName
 import de.schnettler.scrobbler.util.ScrobbleAction
-import de.schnettler.scrobbler.util.asMinSec
 import kotlin.time.Duration
 import kotlin.time.ExperimentalTime
