+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/MenuAction.kt
@@ -1,12 +1,10 @@
-package de.schnettler.scrobbler.util
+package de.schnettler.scrobbler.ui.common.compose
 
 import androidx.annotation.StringRes
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Event
 import androidx.compose.material.icons.rounded.OpenInBrowser
 import androidx.compose.ui.graphics.vector.ImageVector
-import de.schnettler.scrobbler.R
-import de.schnettler.scrobbler.ui.common.compose.UIAction
 
 sealed class MenuAction(@StringRes val label: Int, val icon: ImageVector, val action: UIAction?) {
     object Period : MenuAction(R.string.ic_period, Icons.Rounded.Event, null)
