+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/RefreshableStateViewModel.kt
@@ -1,12 +1,8 @@
-package de.schnettler.scrobbler.viewmodels
+package de.schnettler.scrobbler.ui.common.compose
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.dropbox.android.external.store4.Store
-import de.schnettler.scrobbler.ui.common.compose.RefreshableUiState
-import de.schnettler.scrobbler.ui.common.compose.freshFrom
-import de.schnettler.scrobbler.ui.common.compose.streamFrom
-import de.schnettler.scrobbler.ui.common.compose.updateValue
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.filterNotNull
