+++ ui/common/util/src/main/java/de/schnettler/scrobbler/ui/common/util/DateUtils.kt
@@ -2,8 +2,9 @@ package de.schnettler.scrobbler.ui.common.util
 
 import android.icu.text.CompactDecimalFormat
 import android.icu.text.DateFormat
-import java.util.Date
-import java.util.Locale
+import java.util.*
+import kotlin.time.Duration
+import kotlin.time.ExperimentalTime
 
 fun Long.milliSecondsToDate(): String =
     DateFormat.getDateTimeInstance(DateFormat.MEDIUM, DateFormat.SHORT, Locale.getDefault()).format(
@@ -13,3 +14,9 @@ fun Long.milliSecondsToDate(): String =
 fun Long.abbreviate(): String =
     CompactDecimalFormat.getInstance(Locale.getDefault(), CompactDecimalFormat.CompactStyle.SHORT)
         .format(this)
+
+@ExperimentalTime
+fun Duration.asMinSec() = this.toComponents { min, s, _ ->
+    val padded = s.toString().padStart(2, '0')
+    "$min:$padded"
+}
